+++ packages/trimerge-sync/src/index.ts
@@ -1,7 +1,7 @@
 export * from './differ';
 export * from './TrimergeClient';
 export * from './types';
-export * from './AbstractLocalStore';
+export * from './CoordinatingLocalStore';
 export * from './validateCommits';
 export * from './lib/PromiseQueue';
 export * from './lib/EventChannel';
